fix signed/unsigned comparison Original: 0003a5526a89c8b553c7b2a0f50d5ec0e24a6f20

+++ lalapps/src/inspiral/posterior/SPINspiral/SPINspiral_lal.c
@@ -1243,7 +1243,7 @@ double LALFpFc(LALStatus *status, CoherentGW *waveform, SimInspiralTable *injPar
   
   //      LALSSInjectTimeSeries(status, &chan, &signalvec );
   
-  for ( i = 0; i < signalvec.data->length && i < length; i++ ){
+  for ( i = 0; i < (int)signalvec.data->length && i < length; i++ ){
     
     //printf("%d\t%10.10e\n", i, chan.data->data[i]);
